nix: refactor common.nix
common.nix now contains a higher-order function that returns a module so
that we can pass in inputs and system to the module. We now import it at
the flake level into each system configuration as a module with a helper
so that we don't need to manually merge it inside each configuration.